Add a windows build
Sadly, Travis-CI windows images do not have Python installed (at the
time of this writing, late 2019). So, we ave to install it. Double
sadly, Python is installed into a location that contains the version
number, so in the future if one wants to change the installed Python
version, they will have to update the PATH as well :(